requirements.txt
Django==2.0.3
django-appconf==1.0.2
django-rest-swagger==2.1.2
django-staticfiles==1.2.1
djangorestframework==3.7.7
djangorestframework-simplejwt==3.2.3
whitenoise==3.3.1
gunicorn==19.7.1
PyJWT==1.6.0
PyMySQL==0.8.0
django-storages==1.1.8
boto3==1.5.20
